requirements.txt
openai
waitress
flask
flask-script
gunicorn
python-dotenv
textblob
requests